refactor: extract from action
